// frontend/src/openapi/models/create-chat-room-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateChatRoomRequest
 */
export interface CreateChatRoomRequest {
    /**
     * 
     * @type {number}
     * @memberof CreateChatRoomRequest
     */
    'chat_partner_id': number;
}